requirements.txt
aenum
ansible
attrs
celery
dateutils
dill
pillow
redis
pydicom
requests